refactor(interface): split ui handler code into a module
