+++ src/interface/mod.rs
@@ -1,7 +1,7 @@
 pub mod app;
 mod enums;
 pub mod event;
-pub mod handler;
+pub mod handlers;
 mod popup;
 mod row;
 pub mod stateful_table;
